feat: move order model to own dir and use orders utils
