+++ src/entities/order/model/index.ts
@@ -1,1 +1,2 @@
+export * from './model'
 export * from './utils'
